Split lib.rs (3,042 lines) into smaller focused modules (#519)
Extract FFI boundary types, FFI entry points, and internal business-logic
functions from the monolithic lib.rs into three new focused modules:

- src/ffi_types.rs: JSON request/response structs for the FFI boundary
- src/ffi.rs: #[no_mangle] pub extern "C" FFI entry points
- src/ffi_internal.rs: Internal business-logic functions and unit tests

lib.rs is now 49 lines: module declarations, re-exports, and version init.
All types and functions are re-exported from the crate root so existing
crate::TypeName and neat_ai_discovery::TypeName paths remain unchanged.
