+++ src/ffi.rs
@@ -1,0 +1,1075 @@
+//! FFI entry points — `#[no_mangle] pub extern "C"` functions.
+//!
+//! Every symbol exposed to the Deno FFI layer lives here. Each function
+//! catches panics, converts between C strings and Rust strings, and delegates
+//! to the corresponding `*_internal` function in the crate root.
+
+use crate::ffi_types::*;
+use crate::{log_version_once, streaming};
+
+// ============================================================================
+// Recording — single-call
+// ============================================================================
+
+/// FFI export for recording discovery data
+///
+/// # Safety
+/// This function is unsafe because it deals with raw C strings.
+/// The caller must ensure:
+/// - input_json is a valid null-terminated C string
+/// - The returned pointer is freed using free_discovery_result
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn record_discovery(input_json: *const std::ffi::c_char) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        // Read input C string
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        // Call the Rust function with original string
+        let json_result = match crate::record_discovery_internal(input_str) {
+            Ok(json) => json,
+            Err(e) => {
+                // Properly serialize error message to avoid JSON injection issues
+                let output = RecordDiscoveryOutput {
+                    success: false,
+                    temp_dir: None,
+                    file: None,
+                    error: Some(e.to_string()),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    // Fallback if serialization fails (shouldn't happen)
+                    r#"{"success":false,"error":"Failed to serialize error message"}"#.to_string()
+                })
+            }
+        };
+
+        // Return as C string - this is also protected by catch_unwind
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Streaming Recording API — FFI Entry Points
+// ============================================================================
+// These functions support incremental recording to avoid JavaScript
+// "Invalid string length" errors when serialising large datasets.
+//
+// Usage pattern from TypeScript:
+//   1. start_discovery_session() → returns session_id
+//   2. Loop: append_discovery_records() as data accumulates
+//   3. finish_discovery_session() → finalises Parquet file
+//
+// Benefits:
+//   - Each append call is small enough to serialise (e.g., ~50MB)
+//   - Unlimited total sample size
+//   - Partial data preserved if process crashes
+
+/// Start a new streaming discovery session.
+///
+/// Creates a Parquet file and returns a session ID for subsequent append/finish calls.
+///
+/// Input JSON:
+/// ```json
+/// {
+///   "creature": { ... },
+///   "tempDir": "/path/to/temp/dir"
+/// }
+/// ```
+///
+/// Output JSON:
+/// ```json
+/// {
+///   "success": true,
+///   "sessionId": "uuid-string"
+/// }
+/// ```
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn start_discovery_session(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let input: StartSessionInput = match serde_json::from_str(input_str) {
+            Ok(input) => input,
+            Err(e) => {
+                let output = StartSessionOutput {
+                    success: false,
+                    session_id: None,
+                    error: Some(format!("Failed to parse input JSON: {e}")),
+                };
+                let json = serde_json::to_string(&output).unwrap();
+                return CString::new(json).unwrap().into_raw();
+            }
+        };
+
+        let output = match streaming::start_session(input.creature, input.temp_dir) {
+            Ok(session_id) => StartSessionOutput {
+                success: true,
+                session_id: Some(session_id),
+                error: None,
+            },
+            Err(e) => StartSessionOutput {
+                success: false,
+                session_id: None,
+                error: Some(e.to_string()),
+            },
+        };
+
+        let json = serde_json::to_string(&output).unwrap();
+        CString::new(json).unwrap().into_raw()
+    }))
+    .unwrap_or_else(|panic_info| {
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+/// Append records to an existing streaming session.
+///
+/// Input JSON:
+/// ```json
+/// {
+///   "sessionId": "uuid-string",
+///   "observations": [
+///     {
+///       "obsIndex": 0,
+///       "neuronData": [{ "neuronUuid": "...", "activation": 0.5, "value": 0.4, "errors": [0.1] }],
+///       "inputs": [0.1, 0.2, 0.3]
+///     }
+///   ]
+/// }
+/// ```
+///
+/// Output JSON:
+/// ```json
+/// {
+///   "success": true,
+///   "recordsWritten": 42
+/// }
+/// ```
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn append_discovery_records(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let input: AppendRecordsInput = match serde_json::from_str(input_str) {
+            Ok(input) => input,
+            Err(e) => {
+                let output = AppendRecordsOutput {
+                    success: false,
+                    records_written: None,
+                    error: Some(format!("Failed to parse input JSON: {e}")),
+                };
+                let json = serde_json::to_string(&output).unwrap();
+                return CString::new(json).unwrap().into_raw();
+            }
+        };
+
+        // Convert observations to the internal format
+        let batches: Vec<(u32, Vec<NeuronData>, Vec<f32>)> = input
+            .observations
+            .into_iter()
+            .map(|obs| (obs.obs_index, obs.neuron_data, obs.inputs))
+            .collect();
+
+        let output = match streaming::append_records(&input.session_id, batches) {
+            Ok(records_written) => AppendRecordsOutput {
+                success: true,
+                records_written: Some(records_written),
+                error: None,
+            },
+            Err(e) => AppendRecordsOutput {
+                success: false,
+                records_written: None,
+                error: Some(e.to_string()),
+            },
+        };
+
+        let json = serde_json::to_string(&output).unwrap();
+        CString::new(json).unwrap().into_raw()
+    }))
+    .unwrap_or_else(|panic_info| {
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+/// Finish a streaming session and finalise the Parquet file.
+///
+/// Input JSON:
+/// ```json
+/// {
+///   "sessionId": "uuid-string"
+/// }
+/// ```
+///
+/// Output JSON:
+/// ```json
+/// {
+///   "success": true,
+///   "tempDir": "/path/to/temp/dir",
+///   "file": "discovery_data.parquet",
+///   "totalRecords": 12345
+/// }
+/// ```
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn finish_discovery_session(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let input: FinishSessionInput = match serde_json::from_str(input_str) {
+            Ok(input) => input,
+            Err(e) => {
+                let output = FinishSessionOutput {
+                    success: false,
+                    temp_dir: None,
+                    file: None,
+                    total_records: None,
+                    error: Some(format!("Failed to parse input JSON: {e}")),
+                };
+                let json = serde_json::to_string(&output).unwrap();
+                return CString::new(json).unwrap().into_raw();
+            }
+        };
+
+        let output = match streaming::finish_session(&input.session_id) {
+            Ok((temp_dir, file, total_records)) => FinishSessionOutput {
+                success: true,
+                temp_dir: Some(temp_dir),
+                file: Some(file),
+                total_records: Some(total_records),
+                error: None,
+            },
+            Err(e) => FinishSessionOutput {
+                success: false,
+                temp_dir: None,
+                file: None,
+                total_records: None,
+                error: Some(e.to_string()),
+            },
+        };
+
+        let json = serde_json::to_string(&output).unwrap();
+        CString::new(json).unwrap().into_raw()
+    }))
+    .unwrap_or_else(|panic_info| {
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+/// Cancel a streaming session without finalising.
+///
+/// Use this to clean up if recording fails or is cancelled.
+///
+/// Input JSON:
+/// ```json
+/// {
+///   "sessionId": "uuid-string"
+/// }
+/// ```
+///
+/// Output JSON:
+/// ```json
+/// {
+///   "success": true
+/// }
+/// ```
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn cancel_discovery_session(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let input: CancelSessionInput = match serde_json::from_str(input_str) {
+            Ok(input) => input,
+            Err(e) => {
+                let output = CancelSessionOutput {
+                    success: false,
+                    error: Some(format!("Failed to parse input JSON: {e}")),
+                };
+                let json = serde_json::to_string(&output).unwrap();
+                return CString::new(json).unwrap().into_raw();
+            }
+        };
+
+        let output = match streaming::cancel_session(&input.session_id) {
+            Ok(()) => CancelSessionOutput {
+                success: true,
+                error: None,
+            },
+            Err(e) => CancelSessionOutput {
+                success: false,
+                error: Some(e.to_string()),
+            },
+        };
+
+        let json = serde_json::to_string(&output).unwrap();
+        CString::new(json).unwrap().into_raw()
+    }))
+    .unwrap_or_else(|panic_info| {
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Merge Parquet
+// ============================================================================
+
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn merge_discovery_parquet(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let json_result = match crate::merge_discovery_parquet_internal(input_str) {
+            Ok(json) => json,
+            Err(e) => {
+                let output = MergeParquetOutput {
+                    success: false,
+                    output_file: None,
+                    error: Some(e.to_string()),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    r#"{"success":false,"error":"Failed to serialize error message"}"#.to_string()
+                })
+            }
+        };
+
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Rank focus neurons
+// ============================================================================
+
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn rank_focus_neurons(input_json: *const std::ffi::c_char) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let json_result = match crate::rank_focus_neurons_internal(input_str) {
+            Ok(json) => json,
+            Err(e) => {
+                let output = RankFocusNeuronsOutput {
+                    success: false,
+                    neurons: None,
+                    removal_candidates: None,
+                    constant_neuron_removals: None,
+                    max_output_error: None,
+                    processed_neurons: None,
+                    total_neurons: None,
+                    duration_ms: None,
+                    error: Some(e.to_string()),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    r#"{"success":false,"error":"Failed to serialize output"}"#.to_string()
+                })
+            }
+        };
+
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Analyze parallel
+// ============================================================================
+
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn analyze_parallel(input_json: *const std::ffi::c_char) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let json_result = match crate::analyze_parallel_internal(input_str) {
+            Ok(json) => json,
+            Err(e) => {
+                // Properly serialize error message to avoid JSON injection issues
+                let output = AnalyzeParallelOutput {
+                    success: false,
+                    helpful_synapses: None,
+                    harmful_synapses: None,
+                    synapse_diagnostics: None,
+                    synapse_gpu_used: None,
+                    synapse_metadata: None,
+                    helpful_neurons: None,
+                    synapse_weight_updates: None,
+                    coordinated_structural_candidates: None,
+                    candidate_clusters: None,
+                    neuron_diagnostics: None,
+                    neuron_gpu_used: None,
+                    neuron_metadata: None,
+                    neuron_fingerprints: None,
+                    fingerprint_cache_hits: None,
+                    fingerprint_cache_misses: None,
+                    error: Some(format!("Failed to serialize output: {e}")),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    // Fallback if serialization fails (shouldn't happen)
+                    r#"{"success":false,"error":"Failed to serialize error message"}"#.to_string()
+                })
+            }
+        };
+
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// GPU availability check
+// ============================================================================
+
+#[unsafe(no_mangle)]
+pub extern "C" fn check_gpu_available() -> *mut std::ffi::c_char {
+    use std::ffi::CString;
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let json_result = match crate::check_gpu_available_internal() {
+            Ok(json) => json,
+            Err(e) => {
+                // Properly serialize error message to avoid JSON injection issues
+                let output = CheckGpuOutput {
+                    success: false,
+                    gpu_available: false,
+                    reason: None,
+                    error: Some(format!("Failed to probe GPU: {e}")),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    // Fallback if serialization fails (shouldn't happen)
+                    r#"{"success":false,"gpuAvailable":false,"error":"Failed to serialize error message"}"#.to_string()
+                })
+            }
+        };
+
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"gpuAvailable":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"gpuAvailable\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"gpuAvailable":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Library version
+// ============================================================================
+
+/// FFI export for querying the library version
+///
+/// Returns the version string that was embedded in the binary at compile time.
+/// This allows callers to verify they're using the expected version.
+///
+/// # Safety
+/// The returned pointer must be freed using free_discovery_result
+#[unsafe(no_mangle)]
+pub extern "C" fn get_library_version() -> *mut std::ffi::c_char {
+    use std::ffi::CString;
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let json_result = match crate::get_library_version_internal() {
+            Ok(json) => json,
+            Err(e) => {
+                // Properly serialize error message to avoid JSON injection issues
+                let output = GetVersionOutput {
+                    success: false,
+                    version: String::new(),
+                    error: Some(format!("Failed to get version: {e}")),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    // Fallback if serialization fails (shouldn't happen)
+                    r#"{"success":false,"version":"","error":"Failed to serialize error message"}"#
+                        .to_string()
+                })
+            }
+        };
+
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error =
+                    r#"{"success":false,"version":"","error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"version\":\"\",\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"version":"","error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Read discovery records
+// ============================================================================
+
+/// FFI export for reading discovery records
+///
+/// # Safety
+/// This function is unsafe because it deals with raw C strings.
+/// The caller must ensure:
+/// - input_json is a valid null-terminated C string
+/// - The returned pointer is freed using free_discovery_result
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn read_discovery_records_ffi(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This includes the final CString::new() conversion to catch any panics there
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        // Read input C string
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        // Call the Rust function
+        let json_result = match crate::read_discovery_records(input_str) {
+            Ok(json) => json,
+            Err(e) => {
+                // Properly serialize error message to avoid JSON injection issues
+                let output = ReadDiscoveryOutput {
+                    success: false,
+                    records: None,
+                    error: Some(e.to_string()),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    // Fallback if serialization fails (shouldn't happen)
+                    r#"{"success":false,"error":"Failed to serialize error message"}"#.to_string()
+                })
+            }
+        };
+
+        // Return as C string - this is also protected by catch_unwind
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        // If panic occurred, create a safe error response
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        // This should never fail, but if it does, we return null pointer
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Export visualisation snapshot
+// ============================================================================
+
+/// Export a visualisation snapshot to JSON for debugging with NEAT-AI-Explore.
+///
+/// Input JSON:
+/// ```json
+/// {
+///   "parquetFile": "/path/to/records.parquet",
+///   "creature": { ... },
+///   "outFile": "/path/to/snapshot.json",
+///   "includePerSynapseSeries": true,
+///   "includeReconstructionChecks": true,
+///   "maxObs": null,
+///   "topKWorstSamples": 20
+/// }
+/// ```
+///
+/// Output JSON:
+/// ```json
+/// {
+///   "success": true,
+///   "outFile": "/path/to/snapshot.json",
+///   "stats": { "obsCount": 1000, "neuronCount": 50, "synapseCount": 200, "outputCount": 1 }
+/// }
+/// ```
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn export_visualisation_snapshot(
+    input_json: *const std::ffi::c_char,
+) -> *mut std::ffi::c_char {
+    use std::ffi::{CStr, CString};
+    use std::panic;
+
+    panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        log_version_once();
+
+        let input_str = unsafe {
+            if input_json.is_null() {
+                let error = r#"{"success":false,"error":"Null input pointer"}"#;
+                return CString::new(error).unwrap().into_raw();
+            }
+            match CStr::from_ptr(input_json).to_str() {
+                Ok(s) => s,
+                Err(_) => {
+                    let error = r#"{"success":false,"error":"Invalid UTF-8 in input"}"#;
+                    return CString::new(error).unwrap().into_raw();
+                }
+            }
+        };
+
+        let json_result = match crate::export_visualisation_snapshot_internal(input_str) {
+            Ok(json) => json,
+            Err(e) => {
+                let output = ExportVisualisationSnapshotOutput {
+                    success: false,
+                    out_file: None,
+                    stats: None,
+                    error: Some(e.to_string()),
+                };
+                serde_json::to_string(&output).unwrap_or_else(|_| {
+                    r#"{"success":false,"error":"Failed to serialize error message"}"#.to_string()
+                })
+            }
+        };
+
+        match CString::new(json_result) {
+            Ok(c_string) => c_string.into_raw(),
+            Err(_) => {
+                let error = r#"{"success":false,"error":"Failed to create output string"}"#;
+                CString::new(error).unwrap().into_raw()
+            }
+        }
+    }))
+    .unwrap_or_else(|panic_info| {
+        let msg = if let Some(s) = panic_info.downcast_ref::<&str>() {
+            s.to_string()
+        } else if let Some(s) = panic_info.downcast_ref::<String>() {
+            s.clone()
+        } else {
+            "Unknown panic".to_string()
+        };
+        let error_json = format!(
+            "{{\"success\":false,\"error\":\"Internal panic caught: {}\"}}",
+            msg.replace('\\', "\\\\").replace('"', "\\\"")
+        );
+        CString::new(error_json)
+            .unwrap_or_else(|_| {
+                CString::new(r#"{"success":false,"error":"Failed to create panic error string"}"#)
+                    .unwrap()
+            })
+            .into_raw()
+    })
+}
+
+// ============================================================================
+// Memory management
+// ============================================================================
+
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
+#[unsafe(no_mangle)]
+pub extern "C" fn free_discovery_result(ptr: *mut std::ffi::c_char) {
+    use std::ffi::CString;
+    use std::panic;
+
+    // Catch any panics to prevent unwinding across FFI boundary
+    // This is unlikely to panic, but we protect it anyway for safety
+    let _ = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+        if !ptr.is_null() {
+            unsafe {
+                let _ = CString::from_raw(ptr);
+            }
+        }
+    }));
+}
